reintegrate generate_docs.py into one file
